Fixing Youtube player to use IFrame API
I fixed the player so that it works with the new IFrame API.

Still todo:
-There is currently a problem, where getting a list
of videos in a user's favorites playlist doesn't return
the uploader's channel id in the snippet. Instead it
returns the playlist owner's channel id. I will have to
make a query for each video to determine the owner's
channel id before I can get the favorites for that
channel. This will increate the number of API calls
the application makes.
-Still need to:
-Fix bug: start-index is not supported in V3, need to use pageToken instead
-Fix back end sign in oauth php script
-Make server side tokens go in a separate secret file not to be uploaded to Github
-Make the favorite code work with Youtube API V3
-Allow the user to enter a username for a channel or a channel id.